made types of function IDS elements  and macros assigned to them match
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/epinetics@7 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/src/epinetics_common.h
@@ -21,9 +21,9 @@ struct key
 struct value
 {
   double rate;
-  int function_id;
-  int ego_id;
-  int altar_id;
+  uint16_t function_id;
+  uint32_t ego_id;
+  uint32_t altar_id;
   int strain_id;
   int phylo_id;
 };
